Feat: Add more user-requested tickers to default list
This commit adds a list of specific companies to the default ticker list, as requested by the user.

The company names were mapped to their correct `yfinance` ticker symbols and appended to the `DEFAULT_TICKER_LIST` in `app/ticker_fetcher.py`.

This further expands the global stock universe available for scanning in the default application mode.
