+++ app/main.py
@@ -5,391 +5,30 @@
 import sys
 import pandas as pd
 import webbrowser
-import re
-import configparser
 import os
-from pathlib import Path
 
 from PySide6.QtWidgets import (QApplication, QMainWindow, QWidget, QVBoxLayout,
                              QHBoxLayout, QPushButton, QComboBox, QLabel,
                              QTableWidget, QTableWidgetItem, QProgressBar, QHeaderView,
-                             QMessageBox, QFileDialog, QDialog, QTabWidget, QTextEdit,
-                             QStatusBar, QMenu, QLineEdit, QDialogButtonBox, QListWidget,
-                             QGroupBox, QFormLayout, QSpinBox, QSizePolicy)
-from PySide6.QtCore import QThread, Signal, Qt
+                             QMessageBox, QFileDialog, QStatusBar, QMenu, QLineEdit)
+from PySide6.QtCore import Qt
 from PySide6.QtGui import QAction, QColor, QBrush
 
+# Local Imports
 import screener_engine
 import ticker_fetcher
-from help_texts import HELP_TEXT_DE, HELP_TEXT_EN
-import copy
 import technical_analyzer
-from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg as FigureCanvas
-from matplotlib.figure import Figure
-
-
-# --- Configuration Handling ---
-CONFIG_DIR = Path.home() / ".config" / "rectifex"
-CONFIG_FILE = CONFIG_DIR / "settings.conf"
-
-def load_api_key():
-    if not CONFIG_FILE.exists():
-        return ""
-    config = configparser.ConfigParser()
-    config.read(CONFIG_FILE)
-    return config.get('API', 'key', fallback="")
-
-def save_api_key(key):
-    CONFIG_DIR.mkdir(parents=True, exist_ok=True)
-    config = configparser.ConfigParser()
-    config['API'] = {'key': key}
-    with open(CONFIG_FILE, 'w') as configfile:
-        config.write(configfile)
-
-# --- Dialogs ---
-
-class SettingsDialog(QDialog):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Settings")
-        layout = QVBoxLayout()
-
-        layout.addWidget(QLabel("Enter your Financial Modeling Prep API Key:"))
-        self.api_key_input = QLineEdit()
-        self.api_key_input.setText(load_api_key())
-        layout.addWidget(self.api_key_input)
-
-        info_label = QLabel('<a href="https://site.financialmodelingprep.com/register">Get a free API key from Financial Modeling Prep</a>')
-        info_label.setOpenExternalLinks(True)
-        layout.addWidget(info_label)
-
-        self.button_box = QDialogButtonBox(QDialogButtonBox.Save | QDialogButtonBox.Cancel)
-        self.button_box.accepted.connect(self.accept)
-        self.button_box.rejected.connect(self.reject)
-        layout.addWidget(self.button_box)
-
-        self.setLayout(layout)
-
-    def accept(self):
-        save_api_key(self.api_key_input.text())
-        super().accept()
-
-class ChartDialog(QDialog):
-    """A dialog to display a Matplotlib chart."""
-    def __init__(self, fig: Figure, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Technical Analysis Chart")
-        self.setMinimumSize(900, 700) # Adjusted size
-        self.setModal(False) # Allow interacting with main window
-
-        layout = QVBoxLayout()
-        canvas = FigureCanvas(fig)
-        canvas.setParent(self)
-
-        # Set the size policy to be expanding
-        canvas.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-        canvas.updateGeometry()
-
-        layout.addWidget(canvas)
-
-        self.setLayout(layout)
-
-class ScanWorker(QThread):
-    progress = Signal(tuple); finished = Signal(object)
-    def __init__(self, strategy, tickers, api_key):
-        super().__init__()
-        self.strategy = strategy
-        self.tickers = tickers
-        self.api_key = api_key
-        self._stopped = False
-
-    def run(self):
-        self.finished.emit(screener_engine.run_complete_screener(self.strategy, self.tickers, self.api_key, self.progress, worker=self))
-
-    def stop(self):
-        """Stops the thread."""
-        self._stopped = True
-
-    def is_stopped(self):
-        return self._stopped
-
-class TickerManagerDialog(QDialog):
-    def __init__(self, current_tickers, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Manage Ticker List")
-        self.setMinimumSize(400, 500)
-        self.tickers = current_tickers
-        layout = QVBoxLayout()
-        self.text_edit = QTextEdit()
-        self.text_edit.setPlainText("\n".join(self.tickers))
-        layout.addWidget(self.text_edit)
-        button_layout = QHBoxLayout()
-        self.load_button = QPushButton("Load from File...")
-        self.load_button.clicked.connect(self.load_from_file)
-        button_layout.addWidget(self.load_button)
-        button_layout.addStretch()
-        self.save_button = QPushButton("Save and Close")
-        self.save_button.clicked.connect(self.accept)
-        button_layout.addWidget(self.save_button)
-        self.cancel_button = QPushButton("Cancel")
-        self.cancel_button.clicked.connect(self.reject)
-        button_layout.addWidget(self.cancel_button)
-        layout.addLayout(button_layout)
-        self.setLayout(layout)
-
-    def load_from_file(self):
-        path, _ = QFileDialog.getOpenFileName(self, "Load Tickers", "", "Text Files (*.txt);;CSV Files (*.csv)")
-        if path:
-            try:
-                with open(path, 'r') as f:
-                    content = f.read()
-                    tickers = re.split(r'[,\s;]+', content)
-                    self.text_edit.setPlainText("\n".join(filter(None, tickers)))
-            except Exception as e:
-                QMessageBox.critical(self, "Error", f"Could not load file:\n{e}")
-
-    def get_tickers(self):
-        return [line.strip().upper() for line in self.text_edit.toPlainText().splitlines() if line.strip()]
-
-class HelpDialog(QDialog):
-    def __init__(self, parent=None, error_log=None):
-        super().__init__(parent)
-        self.setWindowTitle("Help & Information")
-        self.setMinimumSize(700, 500)
-        self.tab_widget = QTabWidget()
-        de_tab = QTextEdit(); de_tab.setReadOnly(True); de_tab.setMarkdown(HELP_TEXT_DE)
-        self.tab_widget.addTab(de_tab, "Deutsch")
-        en_tab = QTextEdit(); en_tab.setReadOnly(True); en_tab.setMarkdown(HELP_TEXT_EN)
-        self.tab_widget.addTab(en_tab, "English")
-        if error_log:
-            log_tab = QTextEdit(); log_tab.setReadOnly(True)
-            log_tab.setPlainText("\n".join(error_log))
-            self.tab_widget.addTab(log_tab, "Scan Log")
-            self.tab_widget.setCurrentWidget(log_tab)
-        layout = QVBoxLayout(); layout.addWidget(self.tab_widget); self.setLayout(layout)
-
-class StrategyEditorDialog(QDialog):
-    def __init__(self, strategies, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Strategy Editor")
-        self.setMinimumSize(700, 500)
-
-        self.strategies = copy.deepcopy(strategies)
-        self.current_strategy_name = None
-        self.score_keys = ["Quality_Score", "Value_Score", "Growth_Score", "Momentum_Score", "Yield_Score", "Safety_Score"]
-
-        # --- Main Layout ---
-        main_layout = QVBoxLayout(self)
-        editor_layout = QHBoxLayout()
-        main_layout.addLayout(editor_layout)
-
-        # --- Left Panel: Strategy List ---
-        left_panel = QWidget()
-        left_layout = QVBoxLayout(left_panel)
-        left_layout.addWidget(QLabel("Strategies:"))
-        self.strategy_list_widget = QListWidget()
-        left_layout.addWidget(self.strategy_list_widget)
-
-        list_button_layout = QHBoxLayout()
-        self.new_button = QPushButton("New")
-        self.copy_button = QPushButton("Copy")
-        self.delete_button = QPushButton("Delete")
-        list_button_layout.addWidget(self.new_button)
-        list_button_layout.addWidget(self.copy_button)
-        list_button_layout.addWidget(self.delete_button)
-        left_layout.addLayout(list_button_layout)
-        editor_layout.addWidget(left_panel, 1)
-
-        # --- Right Panel: Editor ---
-        self.editor_groupbox = QGroupBox("Edit Strategy")
-        self.editor_groupbox.setEnabled(False)
-        form_layout = QFormLayout(self.editor_groupbox)
-
-        self.name_input = QLineEdit()
-        form_layout.addRow("Name:", self.name_input)
-
-        self.score_spinboxes = {}
-        for key in self.score_keys:
-            spinbox = QSpinBox()
-            spinbox.setRange(0, 100)
-            spinbox.setSuffix("%")
-            spinbox.valueChanged.connect(self.update_weight_sum)
-            self.score_spinboxes[key] = spinbox
-            form_layout.addRow(key.replace('_', ' ') + ":", spinbox)
-
-        self.sum_label = QLabel("Total: 0%")
-        self.sum_label.setStyleSheet("font-weight: bold;")
-        form_layout.addRow(self.sum_label)
-        editor_layout.addWidget(self.editor_groupbox, 2)
-
-        # --- Dialog Buttons ---
-        self.button_box = QDialogButtonBox(QDialogButtonBox.Save | QDialogButtonBox.Cancel)
-        main_layout.addWidget(self.button_box)
-
-        # --- Connections & Initial State ---
-        self.strategy_list_widget.itemSelectionChanged.connect(self.on_strategy_selection_changed)
-        self.name_input.textChanged.connect(self.on_name_changed)
-        self.new_button.clicked.connect(self.new_strategy)
-        self.copy_button.clicked.connect(self.copy_strategy)
-        self.delete_button.clicked.connect(self.delete_strategy)
-        self.button_box.accepted.connect(self.on_accept)
-        self.button_box.rejected.connect(self.reject)
-
-        self.populate_strategy_list()
-
-    def populate_strategy_list(self, select_name=None):
-        self.strategy_list_widget.blockSignals(True)
-        self.strategy_list_widget.clear()
-        for name in sorted(self.strategies.keys()):
-            self.strategy_list_widget.addItem(name)
-            if name == select_name:
-                self.strategy_list_widget.setCurrentRow(self.strategy_list_widget.count() - 1)
-        self.strategy_list_widget.blockSignals(False)
-        if self.strategy_list_widget.currentItem():
-            self.on_strategy_selection_changed()
-        else:
-            self.editor_groupbox.setEnabled(False)
-
-
-    def on_strategy_selection_changed(self):
-        selected_items = self.strategy_list_widget.selectedItems()
-        if not selected_items:
-            self.current_strategy_name = None
-            self.editor_groupbox.setEnabled(False)
-            return
-
-        new_name = selected_items[0].text()
-        if self.current_strategy_name and self.current_strategy_name != new_name:
-            self.save_current_editor_state()
-
-        self.current_strategy_name = new_name
-        self.load_strategy_into_editor()
-
-    def load_strategy_into_editor(self):
-        if not self.current_strategy_name: return
-
-        strategy_data = self.strategies[self.current_strategy_name]
-        is_predefined = strategy_data.get("predefined", False)
-
-        self.name_input.setText(self.current_strategy_name)
-        self.name_input.setReadOnly(is_predefined)
-        self.delete_button.setEnabled(not is_predefined)
-        self.editor_groupbox.setEnabled(True)
-
-        for key, spinbox in self.score_spinboxes.items():
-            spinbox.blockSignals(True)
-            spinbox.setValue(int(strategy_data.get("weights", {}).get(key, 0) * 100))
-            spinbox.blockSignals(False)
-
-        self.update_weight_sum()
-
-    def save_current_editor_state(self):
-        if not self.current_strategy_name: return
-
-        old_name = self.current_strategy_name
-        new_name = self.name_input.text()
-
-        if not new_name:
-             # User cleared the name field, revert to old name to avoid issues
-            self.name_input.setText(old_name)
-            new_name = old_name
-
-        strategy_data = self.strategies.pop(old_name)
-        strategy_data["weights"] = {key: spinbox.value() / 100.0 for key, spinbox in self.score_spinboxes.items()}
-        self.strategies[new_name] = strategy_data
-        self.current_strategy_name = new_name
-
-
-    def on_name_changed(self, new_name):
-        if not self.current_strategy_name: return
-
-        selected_items = self.strategy_list_widget.selectedItems()
-        if not selected_items: return
-
-        # Prevent renaming to an existing name
-        if new_name != self.current_strategy_name and new_name in self.strategies:
-            self.name_input.setStyleSheet("color: red;")
-        else:
-            self.name_input.setStyleSheet("")
-
-        selected_items[0].setText(new_name)
-
-
-    def update_weight_sum(self):
-        total_weight = sum(spinbox.value() for spinbox in self.score_spinboxes.values())
-        self.sum_label.setText(f"Total: {total_weight}%")
-        if total_weight != 100:
-            self.sum_label.setStyleSheet("color: red; font-weight: bold;")
-        else:
-            self.sum_label.setStyleSheet("color: green; font-weight: bold;")
-
-    def new_strategy(self):
-        self.save_current_editor_state()
-        base_name = "New Strategy"
-        name = base_name
-        i = 1
-        while name in self.strategies:
-            name = f"{base_name} {i}"
-            i += 1
-
-        self.strategies[name] = {"weights": {key: 0 for key in self.score_keys}, "predefined": False}
-        self.populate_strategy_list(select_name=name)
-
-    def copy_strategy(self):
-        if not self.current_strategy_name: return
-        self.save_current_editor_state()
-
-        base_name = f"{self.current_strategy_name} (Copy)"
-        name = base_name
-        i = 1
-        while name in self.strategies:
-            name = f"{base_name} {i}"
-            i += 1
-
-        original_strategy = self.strategies[self.current_strategy_name]
-        self.strategies[name] = {
-            "weights": original_strategy["weights"].copy(),
-            "predefined": False
-        }
-        self.populate_strategy_list(select_name=name)
-
-    def delete_strategy(self):
-        if not self.current_strategy_name or self.strategies[self.current_strategy_name].get("predefined", False):
-            return
-
-        reply = QMessageBox.question(self, "Confirm Deletion", f"Are you sure you want to delete '{self.current_strategy_name}'?")
-        if reply == QMessageBox.Yes:
-            del self.strategies[self.current_strategy_name]
-            self.current_strategy_name = None
-            self.populate_strategy_list()
-
-    def on_accept(self):
-        self.save_current_editor_state()
-
-        invalid_strategies = []
-        for name, data in self.strategies.items():
-            total_weight = sum(data.get("weights", {}).values())
-            if not (0.999 < total_weight < 1.001):
-                invalid_strategies.append(name)
-
-        if invalid_strategies:
-            QMessageBox.warning(self, "Invalid Weights", "The following strategies do not have weights summing to 100%:\n\n" + "\n".join(invalid_strategies))
-            return
-
-        # Check for duplicate names
-        if len(self.strategies.keys()) != len(set(self.strategies.keys())):
-             QMessageBox.warning(self, "Duplicate Names", "Strategy names must be unique.")
-             return
-
-        self.accept()
-
-    def get_updated_strategies(self):
-        return self.strategies
+from scan_worker import ScanWorker
+from ui.settings_dialog import SettingsDialog, load_api_key
+from ui.chart_dialog import ChartDialog
+from ui.ticker_manager_dialog import TickerManagerDialog
+from ui.help_dialog import HelpDialog
+from ui.strategy_editor_dialog import StrategyEditorDialog
 
 class MainWindow(QMainWindow):
     def __init__(self):
         super().__init__(); self.setWindowTitle("Rectifex - Global Stock Screener"); self.setGeometry(100, 100, 1200, 800); self.result_df = None
-        self.current_tickers = screener_engine.get_default_tickers()
+        self.current_tickers = [] # Start with an empty list
         self.scan_errors = []
         self.api_key = load_api_key()
         self.chart_cache = {}
@@ -412,8 +51,7 @@ class MainWindow(QMainWindow):
         self.ticker_source_label = QLabel("Ticker Source:")
         self.ticker_source_combo = QComboBox()
         self.ticker_source_combo.addItems([
-            "Default List", "DAX", "MDAX", "SDAX", "TecDAX",
-            "S&P 500", "Nasdaq 100", "Dow Jones", "Custom List"
+            "Default Global List", "Custom List"
         ])
         self.ticker_source_combo.currentIndexChanged.connect(self.on_ticker_source_changed)
         self.manage_tickers_button = QPushButton("Manage Custom List")
@@ -518,47 +156,15 @@ class MainWindow(QMainWindow):
 
     def start_scan(self):
         ticker_source = self.ticker_source_combo.currentText()
-        self.current_tickers = []
-        new_tickers = None
 
-        QApplication.setOverrideCursor(Qt.WaitCursor)
-        self.statusBar().showMessage(f"Fetching tickers for {ticker_source}...")
-        QApplication.processEvents()
-
-        try:
-            if ticker_source == "Default List":
-                new_tickers = screener_engine.get_default_tickers()
-            elif ticker_source == "Custom List":
-                # No fetch needed, current_tickers is managed by dialog
-                new_tickers = self.current_tickers
-            elif ticker_source == "DAX":
-                new_tickers = ticker_fetcher.get_dax_tickers()
-            elif ticker_source == "MDAX":
-                new_tickers = ticker_fetcher.get_mdax_tickers()
-            elif ticker_source == "SDAX":
-                new_tickers = ticker_fetcher.get_sdax_tickers()
-            elif ticker_source == "TecDAX":
-                new_tickers = ticker_fetcher.get_tecdax_tickers()
-            else: # US Indices that require an API key
-                if not self.api_key:
-                    QMessageBox.critical(self, "API Key Required", f"An FMP API key is required to fetch tickers from the {ticker_source} index.\n\nPlease add one in Settings.")
-                    new_tickers = None # Ensure we don't proceed
-                else:
-                    new_tickers = screener_engine.get_tickers_from_index(ticker_source, self.api_key)
-
-            if new_tickers is None and ticker_source not in ["Custom List"]:
-                 QMessageBox.critical(self, "Error", f"Failed to fetch ticker list for {ticker_source}.\nThis might be a network issue or an invalid API key.")
-                 self.statusBar().showMessage(f"Failed to fetch tickers for {ticker_source}.", 8000)
-            else:
-                self.current_tickers = new_tickers
-
-        finally:
-            QApplication.restoreOverrideCursor()
-            self.statusBar().clearMessage()
+        if ticker_source == "Default Global List":
+            self.current_tickers = ticker_fetcher.get_default_tickers()
+        elif ticker_source == "Custom List":
+            # self.current_tickers is already updated by the dialog, so no action needed here.
+            pass
 
         if not self.current_tickers:
-            if ticker_source != "Custom List": # Avoid warning for an empty custom list on startup
-                QMessageBox.warning(self, "No Tickers", f"The ticker list for '{ticker_source}' is empty or could not be loaded.")
+            QMessageBox.warning(self, "No Tickers", f"The selected ticker list '{ticker_source}' is empty.")
             return
 
         # --- Start the Scan ---
@@ -604,6 +210,17 @@ class MainWindow(QMainWindow):
         self.scan_errors = summary.get('failed_list', [])
         status_msg = f"Scan finished. {summary.get('final_count', 0)} of {summary.get('total_tickers', 0)} stocks loaded. {summary.get('failed_count', 0)} tickers failed."
         self.statusBar().showMessage(status_msg, 10000)
+
+        # If all tickers failed, show a helpful message to the user.
+        if summary.get('final_count', 0) == 0 and summary.get('failed_count', 0) > 0:
+            error_preview = "\n".join(self.scan_errors[:3]) # Show first 3 errors
+            QMessageBox.warning(self, "Scan Completed with Errors",
+                                f"The scan finished, but all {summary.get('failed_count', 0)} tickers failed to load.\n\n"
+                                "This is often caused by an invalid or expired API key, or network connectivity issues.\n\n"
+                                f"Here are some of the errors:\n{error_preview}...\n\n"
+                                "Please check your API key in Settings or view the full log in the 'Help' menu.")
+            return
+
         if df.empty: return
         self.result_df = df
         self.save_csv_button.setEnabled(True)
