+++ app/requirements.txt
@@ -1,8 +1,43 @@
-pandas
-yfinance
-numpy
-PySide6
-requests
-matplotlib
-mplfinance
-lxml
+beautifulsoup4==4.13.5
+certifi==2025.8.3
+cffi==2.0.0
+charset-normalizer==3.4.3
+contourpy==1.3.3
+curl_cffi==0.13.0
+cycler==0.12.1
+fonttools==4.59.2
+frozendict==2.4.6
+greenlet==3.2.3
+idna==3.10
+kiwisolver==1.4.9
+lxml==6.0.1
+matplotlib==3.10.6
+mplfinance==0.12.10b0
+multitasking==0.0.12
+numpy==2.3.3
+packaging==25.0
+pandas==2.3.2
+peewee==3.18.2
+pillow==11.3.0
+platformdirs==4.4.0
+playwright==1.54.0
+protobuf==6.32.1
+pycparser==2.23
+pyee==13.0.0
+pyparsing==3.2.4
+PySide6==6.9.2
+PySide6_Addons==6.9.2
+PySide6_Essentials==6.9.2
+python-dateutil==2.9.0.post0
+pytz==2025.2
+requests==2.32.5
+shiboken6==6.9.2
+six==1.17.0
+soupsieve==2.8
+typing_extensions==4.14.1
+tzdata==2025.2
+urllib3==2.5.0
+websockets==15.0.1
+yfinance==0.2.65
+pytest
+pytest-mock
