+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+testpaths = app
+python_files = test_*.py
+addopts = -p no:warnings --ignore=app/test_ui.py
